small refactor.
