+++ src/services/vuestic-ui/components/va-icon.ts
